Rewrite the Heimdal plugin tests in C
Add a new Perl program to generate a C version of the password
test data from the JSON source, and run that program during
autogen so that we can ship the source in the distribution and
not require JSON support for the basic test suite.  Add a new
shell script to generate the necessary krb5.conf file.

Rewrite the Heimdal plugin test in pure C, using the generated C
data and calling the shell script to create the necessary krb5.conf
file.