Move krb5.conf configuration into the plugin
Load the configuration inside the plugin when we initialize it,
and pass in a Kerberos context to the plugin so that this is
possible.  Obtain or pass in an existing Kerberos context in the
places where we weren't already doing so.

This loses some more detailed error reporting, particularly around
non-existent configured dictionaries, which will be restored later
by using the Kerberos error message.