Add password history implementation for Heimdal
A password history implementation for Heimdal is now included.  This
is a separate Perl program, heimdal-history, that stacks with the
external program implementation of strength checking.  It is not
available in the form of a plugin, only as a Heimdal external password
quality check.  (MIT Kerberos provides its own password history
mechanism.)  This program has more extensive Perl module dependencies
than the other programs in this distribution.