Skip Perl strictness testing if module prereqs are missing
Modify the standard Perl strictness test, which also checks Perl
scripts for syntax errors, to support a list of prerequisite
modules.  Skip the test if any of those modules can't be loaded,
since they'll otherwise cause failures.  This makes the testing
more robust given that we have some scripts that require a bunch
of Perl modules not needed by the main package.