Move configuration instructions to man pages
Create a new krb5-strength man page that gets the configuration
instructions for the plugins (with the proper path substituted in
by the Makefile) and move other configuration details to the
heimdal-strength and heimdal-history man pages.  Duplicate the
documentation for cracklib_maxlen in the heimdal-strength man page.

This will make it easier to automate generation of the README file,
since it will now require less complex formatting.