Handle errors in heimdal-strength tests better
In Travis CI, the heimdal-strength test was failing with a broken
pipe.  Apparently the timing was such that the child process would
exit with an error before IPC::Run could try to write the input,
and then IPC::Run would die with a broken pipe.  Work around this
by allowing a flag to be passed to run_heimdal_strength saying whether
to expect an error, and don't send input in that case.